Project: use `RemoteRepository` to define `default_branch`
Currently, we are using `Project.default_branch` when it's defined and falling
back to VCS's default branch. However, with the massive rename from `master` to
`main` in GitHub, a lot of project are pointing to the wrong default branch.

Since we are storing the default branch of the GitHub repository under
`RemoteRepository.default_branch` and we are refreshing this list more
frequently now, we can use it before using VCS's default branch as fallback.

Closes #7907